+++ src/hooks/useMnemonic.tsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 
 export const useMnemonic = (
   initialMnemonic: string[] = Array(12).fill(''),
@@ -18,25 +18,4 @@ export const useMnemonic = (
     setMnemonic(mnemonicCopy);
   };
   return [mnemonic, setMnemonicWord];
-};
-
-export const useFocus = (
-  numberOfInputs: number,
-  onEnd?: () => void,
-): [React.RefObject<HTMLIonInputElement>[], (i: number) => void] => {
-  const refs: React.RefObject<HTMLIonInputElement>[] = [];
-
-  for (let i = 0; i < numberOfInputs; i++) {
-    refs[i] = useRef(null);
-  }
-
-  const setFocus = (i: number) => {
-    if (onEnd && i === numberOfInputs) onEnd();
-
-    if (refs[i]) {
-      refs[i].current?.setFocus();
-    }
-  };
-
-  return [refs, setFocus];
 };
